Добавить kafka container, kafka producer

+++ src/main/java/ru/acherm/astonhw2/service/UserServiceImpl.java
@@ -1,9 +1,13 @@
 package ru.acherm.astonhw2.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 import ru.acherm.astonhw2.dao.UserRepository;
 import ru.acherm.astonhw2.dto.UserDto;
+import ru.acherm.astonhw2.event.UserCreatedEvent;
+import ru.acherm.astonhw2.event.UserDeletedEvent;
+import ru.acherm.astonhw2.event.UserEvent;
 import ru.acherm.astonhw2.mapper.UserMapper;
 import ru.acherm.astonhw2.model.User;
 
@@ -13,16 +17,25 @@ import java.util.Optional;
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final UserMapper userMapper;
+    private final KafkaTemplate<String, UserEvent> kafkaTemplate;
 
     @Autowired
-    public UserServiceImpl(UserRepository userRepository, UserMapper userMapper) {
+    public UserServiceImpl(UserRepository userRepository,
+                           UserMapper userMapper,
+                           KafkaTemplate<String, UserEvent> kafkaTemplate) {
         this.userRepository = userRepository;
         this.userMapper = userMapper;
+        this.kafkaTemplate = kafkaTemplate;
     }
 
     @Override
     public UserDto create(UserDto user) {
-        return userMapper.toDto(userRepository.save(userMapper.toEntity(user)));
+        UserDto createdDto = userMapper.toDto(userRepository.save(userMapper.toEntity(user)));
+        UserCreatedEvent createdEvent = userMapper.toCreatedEvent(createdDto);
+
+        kafkaTemplate.send("user-created-event-topic",
+                createdEvent.getId().toString(), createdEvent);
+        return createdDto;
     }
 
     @Override
@@ -50,6 +63,13 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delete(long id) {
-        userRepository.deleteById(id);
+        Optional<User> optUser = userRepository.findById(id);
+
+        if (optUser.isPresent()) {
+            UserDeletedEvent deletedEvent = userMapper.toDeletedEvent(optUser.get());
+            userRepository.deleteById(id);
+            kafkaTemplate.send("user-deleted-event-topic",
+                    deletedEvent.getId().toString(), deletedEvent);
+        }
     }
 }
